refactored masstive controller into smaller  categorized controllers
