+++ Sources/SteamPress/Extensions/Request+Utils.swift
@@ -1,0 +1,11 @@
+import Vapor
+
+extension Request {
+    func getURIWithHTTPSIfReverseProxy() -> URI {
+        if self.headers["X-Forwarded-Proto"] == "https" {
+            return URI(scheme: "https", userInfo: self.uri.userInfo, hostname: self.uri.hostname, port: nil, path: self.uri.path, query: self.uri.query, fragment: self.uri.fragment)
+        }
+
+        return self.uri
+    }
+}
