+++ Sources/SteamPress/Validators/Password.swift
@@ -1,7 +1,6 @@
 import Validation
 
 struct PasswordValidator: Validator {
-
     func validate(_ input: String) throws {
         // Check length
         if input.count < 8 {
